+++ changelog.md
@@ -11,6 +11,9 @@
     - [#341](https://github.com/bitemyapp/esqueleto/pull/341/)
         - Add functions for `NULLS FIRST` and `NULLS LAST` in the Postgresql
           module
+- @JoelMcCracken
+    - [#354](https://github.com/bitemyapp/esqueleto/pull/354), [#417](https://github.com/bitemyapp/esqueleto/pull/417)
+      - Add `withMaterialized`, `withNotMaterialized` to the PostgreSQL module
 
 3.5.13.2
 ========